test(copyright): split detector pipeline test buckets
