+++ src/copyright/detector/tests.rs
@@ -156,21 +156,6 @@ fn test_swift_convention_c_signatures_do_not_produce_copyrights_or_holders() {
 }
 
 #[test]
-fn test_author_prefix_dedup_keeps_short_email_list() {
-    let input = "Author(s): gthomas, sorin@netappi.com\nContributors: gthomas, sorin@netappi.com, andrew.lunn@ascom.ch\n";
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-    let vals: Vec<&str> = authors.iter().map(|a| a.author.as_str()).collect();
-    assert!(
-        vals.contains(&"gthomas, sorin@netappi.com"),
-        "authors: {vals:?}"
-    );
-    assert!(
-        vals.contains(&"gthomas, sorin@netappi.com, andrew.lunn@ascom.ch"),
-        "authors: {vals:?}"
-    );
-}
-
-#[test]
 fn test_added_copyright_year_for_line_is_extracted() {
     let input = "Added the Copyright year (2020) for A11yance";
 
@@ -190,30 +175,17 @@ fn test_added_copyright_year_for_line_is_extracted() {
 }
 
 #[test]
-fn test_was_developed_by_multiline_author_is_extracted() {
-    let input = "1. GOST R 34.11-2012 was developed by the Center for Information\nProtection and Special Communications of the Federal Security\nService of the Russian Federation with participation of the Open\n";
-
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
+fn test_author_prefix_dedup_keeps_short_email_list() {
+    let input = "Author(s): gthomas, sorin@netappi.com\nContributors: gthomas, sorin@netappi.com, andrew.lunn@ascom.ch\n";
+    let (_c, _h, authors) = detect_copyrights_from_text(input);
+    let vals: Vec<&str> = authors.iter().map(|a| a.author.as_str()).collect();
     assert!(
-            authors.iter().any(|a| {
-                a.author
-                    == "the Center for Information Protection and Special Communications of the Federal Security Service of the Russian Federation"
-            }),
-            "authors: {:?}",
-            authors.iter().map(|a| &a.author).collect::<Vec<_>>()
-        );
-}
-
-#[test]
-fn test_written_by_author_email_for_project_is_extracted() {
-    let input = "Written by Andy Polyakov <appro@openssl.org> for the OpenSSL\nproject.";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
+        vals.contains(&"gthomas, sorin@netappi.com"),
+        "authors: {vals:?}"
+    );
     assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Andy Polyakov <appro@openssl.org>"),
-        "authors: {:?}",
-        authors.iter().map(|a| &a.author).collect::<Vec<_>>()
+        vals.contains(&"gthomas, sorin@netappi.com, andrew.lunn@ascom.ch"),
+        "authors: {vals:?}"
     );
 }
 
@@ -262,66 +234,6 @@ fn test_boost_html_holder_drops_symbol_table_run_junk() {
 }
 
 #[test]
-fn test_prose_snippet_does_not_report_laboriously_took_the_trouble_as_author() {
-    let input = concat!(
-        "<para>the authors laboriously took the trouble of searching for workarounds ",
-        "to make these compilers happy</para>",
-    );
-
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {:?}", authors);
-}
-
-#[test]
-fn test_developed_by_sentence_author_is_extracted() {
-    let input = "developed by the U.S. Government. BAE Systems is enhancing and supporting the SMP";
-
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "the U.S. Government. BAE Systems"),
-        "authors: {:?}",
-        authors.iter().map(|a| &a.author).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_developed_by_phrase_author_is_extracted() {
-    let input = "to acknowledge that it was\n\
-      developed by the National Center for Supercomputing Applications at the University of Illinois at Urbana-Champaign and to credit the\n\
-      contributors.";
-
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-    assert!(
-            authors.iter().any(|a| {
-                a.author
-                    == "the National Center for Supercomputing Applications at the University of Illinois at Urbana-Champaign"
-            }),
-            "authors: {:?}",
-            authors.iter().map(|a| &a.author).collect::<Vec<_>>()
-        );
-}
-
-#[test]
-fn test_modified_portion_developed_by_author_with_url_is_extracted() {
-    let input = concat!(
-        "# This product contains a modified portion of 'Flask App Builder' developed by Daniel Vaz Gaspar.\n",
-        "# (https://github.com/dpgaspar/Flask-AppBuilder).\n",
-    );
-
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-    assert!(
-        authors.iter().any(
-            |a| a.author == "Daniel Vaz Gaspar. (https://github.com/dpgaspar/Flask-AppBuilder)"
-        ),
-        "authors: {:?}",
-        authors.iter().map(|a| &a.author).collect::<Vec<_>>()
-    );
-}
-
-#[test]
 fn test_current_year_placeholder_copyright_holder_detected() {
     let input = "Copyright 2016- CURRENT_YEAR The Apache Software Foundation";
 
@@ -341,79 +253,6 @@ fn test_current_year_placeholder_copyright_holder_detected() {
             .any(|h| h.holder == "The Apache Software Foundation"),
         "holders: {:?}",
         holders.iter().map(|h| &h.holder).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_author_colon_block_stops_at_status_and_devices_metadata() {
-    let input = "Author: ds\nStatus: works in immediate mode\nDevices: [standard] parallel port\n";
-
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.is_empty(),
-        "authors: {:?}",
-        authors.iter().map(|a| &a.author).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_author_colon_block_keeps_named_author_without_devices_tail() {
-    let input =
-        "Author: Pablo Mejia <pablo.mejia@cctechnol.com>\nDevices: [Access I/O] PC-104 AIO12-8\n";
-
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Pablo Mejia <pablo.mejia@cctechnol.com>"),
-        "authors: {:?}",
-        authors.iter().map(|a| &a.author).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_written_by_comma_and_copyright_keeps_parenthesized_email_author() {
-    let input =
-        "written by Philip Hazel, and copyright\nby the University of Cambridge, England.\n";
-
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Philip Hazel"),
-        "authors: {:?}",
-        authors.iter().map(|a| &a.author).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_with_additional_hacking_by_keeps_parenthesized_email_author() {
-    let input = "With additional hacking by Jeffrey Kuskin (jsk@mojave.stanford.edu)\n";
-
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Jeffrey Kuskin (jsk@mojave.stanford.edu)"),
-        "authors: {:?}",
-        authors.iter().map(|a| &a.author).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_adapted_written_by_keeps_parenthesized_email_author() {
-    let input = "Adapted from baycom.c driver written by Thomas Sailer (sailer@ife.ee.ethz.ch)\n";
-
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Thomas Sailer (sailer@ife.ee.ethz.ch)"),
-        "authors: {:?}",
-        authors.iter().map(|a| &a.author).collect::<Vec<_>>()
     );
 }
 
@@ -623,40 +462,6 @@ fn test_copyright_span_does_not_absorb_following_lint_directive_line() {
     assert!(
         !values.iter().any(|c| c.contains("@lint-ignore-every")),
         "copyrights: {values:#?}"
-    );
-}
-
-#[test]
-fn test_multiline_written_and_maintained_by_block_extracts_individual_authors() {
-    let input = concat!(
-        "GNU tar, heavily based on John Gilmore's public domain version of tar,\n",
-        "was originally written by Graham Todd.\n",
-        "It is now maintained by Sergey Poznyakoff.\n",
-        "This package is maintained for Debian by Janos Lenart <ocsi@debian.org>.\n",
-    );
-
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-    let authors: Vec<String> = authors.into_iter().map(|a| a.author).collect();
-
-    assert!(
-        authors.iter().any(|a| a == "Graham Todd"),
-        "authors: {authors:#?}"
-    );
-    assert!(
-        authors.iter().any(|a| a == "Sergey Poznyakoff"),
-        "authors: {authors:#?}"
-    );
-    assert!(
-        authors
-            .iter()
-            .any(|a| a == "Janos Lenart <ocsi@debian.org>"),
-        "authors: {authors:#?}"
-    );
-    assert!(
-        !authors
-            .iter()
-            .any(|a| a.contains("GNU tar, heavily based on")),
-        "authors: {authors:#?}"
     );
 }
 
@@ -1516,136 +1321,6 @@ fn test_modified_by_lines_extract_real_authors_without_cpusets_prose_false_posit
 }
 
 #[test]
-fn test_rst_field_author_and_maintainer_extracts_single_author() {
-    let input = ":License:\t\tGPLv2\n:Author & Maintainer:\tMiguel Ojeda <ojeda@kernel.org>\n:Date:\t\t\t2006-10-27\n";
-
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-    let values: Vec<&str> = authors
-        .iter()
-        .map(|author| author.author.as_str())
-        .collect();
-    assert!(
-        values.contains(&"Miguel Ojeda <ojeda@kernel.org>"),
-        "authors: {values:?}"
-    );
-}
-
-#[test]
-fn test_maintainers_label_extracts_author_and_trims_gitrepo_suffix() {
-    let input = "Maintainers Tianon Gravi <admwiggin@gmail.com> (@tianon) GitRepo https://github.com/tianon/docker-bash.git\n";
-
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-    let values: Vec<&str> = authors
-        .iter()
-        .map(|author| author.author.as_str())
-        .collect();
-    assert!(
-        values.contains(&"Tianon Gravi <admwiggin@gmail.com> (@tianon)"),
-        "authors: {values:?}"
-    );
-    assert!(
-        !values.iter().any(|value| value.contains("GitRepo")),
-        "authors: {values:?}"
-    );
-}
-
-#[test]
-fn test_maintainer_comment_extracts_author_without_label() {
-    let input = "# Maintainer: Sébastien Luttringer <seblu@archlinux.org>\n";
-
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-    let values: Vec<&str> = authors
-        .iter()
-        .map(|author| author.author.as_str())
-        .collect();
-
-    assert!(
-        values.contains(&"Sébastien Luttringer <seblu@archlinux.org>"),
-        "authors: {values:?}"
-    );
-    assert!(
-        !values.iter().any(|value| value.contains("Maintainer")),
-        "authors: {values:?}"
-    );
-}
-
-#[test]
-fn test_maintainers_label_without_email_does_not_extract_author() {
-    let input = "Maintainers the Docker Community\n";
-
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_dash_bullet_changelog_lines_extract_individual_authors() {
-    let input = "- Written by Mydraal <vulpyne@vulpyne.net>\n- Updated by Adam Sulmicki <adam@cfar.umd.edu>\n- Updated by Jeremy M. Dolan <jmd@turbogeek.org> 2001/01/28 10:15:59\n- Added to by Crutcher Dunnavant <crutcher+kernel@datastacks.com>\n";
-
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-    let values: Vec<&str> = authors
-        .iter()
-        .map(|author| author.author.as_str())
-        .collect();
-    assert!(
-        values.contains(&"Mydraal <vulpyne@vulpyne.net>"),
-        "authors: {values:?}"
-    );
-    assert!(
-        values.contains(&"Adam Sulmicki <adam@cfar.umd.edu>"),
-        "authors: {values:?}"
-    );
-    assert!(
-        values.contains(&"Jeremy M. Dolan <jmd@turbogeek.org>"),
-        "authors: {values:?}"
-    );
-    assert!(
-        values.contains(&"Crutcher Dunnavant <crutcher+kernel@datastacks.com>"),
-        "authors: {values:?}"
-    );
-    assert!(
-        !values
-            .iter()
-            .any(|value| value.contains("Updated by Adam Sulmicki")),
-        "authors: {values:?}"
-    );
-}
-
-#[test]
-fn test_author_colon_dash_bullet_hwmon_roster_extracts_individual_authors() {
-    let input = "Authors:\n\t- Mark M. Hoffman <mhoffman@lightlink.com>\n\t- Ported to 2.6 by Eric J. Bowersox <ericb@aspsys.com>\n\t- Adapted to 2.6.20 by Carsten Emde <ce@osadl.org>\n\t- Modified for mainline integration by Hans J. Koch <hjk@hansjkoch.de>\n";
-
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-    let values: Vec<&str> = authors
-        .iter()
-        .map(|author| author.author.as_str())
-        .collect();
-    assert!(
-        values.contains(&"Mark M. Hoffman <mhoffman@lightlink.com>"),
-        "authors: {values:?}"
-    );
-    assert!(
-        values.contains(&"Eric J. Bowersox <ericb@aspsys.com>"),
-        "authors: {values:?}"
-    );
-    assert!(
-        values.contains(&"Carsten Emde <ce@osadl.org>"),
-        "authors: {values:?}"
-    );
-    assert!(
-        values.contains(&"Hans J. Koch <hjk@hansjkoch.de>"),
-        "authors: {values:?}"
-    );
-}
-
-#[test]
-fn test_passive_written_phrase_does_not_create_abi_author_false_positive() {
-    let input = "Description:\tWhen read, this file returns general data like firmware version.\n\t\tWhen written, the device can be reset.\n\t\tBefore reading this file, control has to be written to select\n\t\twhich profile to read.\n";
-
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
 fn test_copyright_year_range_only_detected() {
     let input = "Copyright (c) 1995-1999.";
     let (c, _h, _a) = detect_copyrights_from_text(input);
@@ -1959,103 +1634,6 @@ fn test_detect_multiline_copyright() {
     let text = "Copyright 2024\n  Acme Corporation\n  All rights reserved.";
     let (c, _h, _a) = detect_copyrights_from_text(text);
     assert!(!c.is_empty(), "Should detect multiline copyright");
-}
-
-#[test]
-fn test_detect_author() {
-    let (c, h, a) = detect_copyrights_from_text("Written by John Doe");
-    // "Written" is tagged Auth2, triggering author span extraction.
-    assert!(c.is_empty(), "Should not detect copyright");
-    assert!(h.is_empty(), "Should not detect holder");
-    assert_eq!(a.len(), 1, "Should detect one author, got: {:?}", a);
-    assert_eq!(a[0].author, "John Doe");
-    assert_eq!(a[0].start_line, LineNumber::ONE);
-    assert_eq!(a[0].end_line, LineNumber::ONE);
-}
-
-#[test]
-fn test_detect_author_from_xml_author_attribute() {
-    let text = r#"<note author="Vinnie Falco">C++11 is the minimum requirement.</note>"#;
-    let (c, h, a) = detect_copyrights_from_text(text);
-
-    assert!(c.is_empty(), "Should not detect copyright");
-    assert!(h.is_empty(), "Should not detect holder");
-    assert_eq!(a.len(), 1, "Should detect one author, got: {:?}", a);
-    assert_eq!(a[0].author, "Vinnie Falco");
-    assert_eq!(a[0].start_line, LineNumber::ONE);
-    assert_eq!(a[0].end_line, LineNumber::ONE);
-}
-
-#[test]
-fn test_detect_author_from_xml_author_attribute_without_note_body_noise() {
-    let text = r#"<note author="Chris Kohlhoff">
-This compiler does not support enable_if, which is needed by the library.
-</note>"#;
-    let (_c, _h, a) = detect_copyrights_from_text(text);
-
-    assert_eq!(a.len(), 1, "Should detect one author, got: {:?}", a);
-    assert_eq!(a[0].author, "Chris Kohlhoff");
-}
-
-#[test]
-fn test_detect_author_from_xml_author_attribute_decodes_entities() {
-    let text = r#"<note author="Joaqu&#237;n M L&#243;pez Mu&#241;oz">Compiler bug.</note>"#;
-    let (_c, _h, a) = detect_copyrights_from_text(text);
-
-    assert_eq!(a.len(), 1, "Should detect one author, got: {:?}", a);
-    assert_eq!(a[0].author, "Joaquín M López Muñoz");
-}
-
-#[test]
-fn test_detect_author_from_repeated_xml_author_attributes_keeps_multiple_occurrences() {
-    let text = r#"<mark-expected-failures>
-<note author="Aleksey Gurtovoy" refid="4"/>
-<note author="Aleksey Gurtovoy" refid="19"/>
-</mark-expected-failures>"#;
-    let (_c, _h, a) = detect_copyrights_from_text(text);
-
-    let matching: Vec<_> = a
-        .iter()
-        .filter(|ad| ad.author == "Aleksey Gurtovoy")
-        .collect();
-    assert_eq!(matching.len(), 2, "authors: {a:#?}");
-    assert_eq!(matching[0].start_line, LineNumber::new(2).expect("valid"));
-    assert_eq!(matching[1].start_line, LineNumber::new(3).expect("valid"));
-}
-
-#[test]
-fn test_detect_author_from_xml_author_attribute_splits_obvious_multi_name_lists() {
-    let text = r#"<note author="Robert Ramey,Roland Schwarz" date="16 Feb 07" refid="19"/>"#;
-    let (_c, _h, a) = detect_copyrights_from_text(text);
-
-    let names: Vec<&str> = a.iter().map(|ad| ad.author.as_str()).collect();
-    assert!(names.contains(&"Robert Ramey"), "authors: {names:?}");
-    assert!(names.contains(&"Roland Schwarz"), "authors: {names:?}");
-    assert_eq!(names.len(), 2, "authors: {names:?}");
-}
-
-#[test]
-fn test_detect_docbook_html_authorgroup_authors() {
-    let text = r#"<div class="authorgroup">
-<div class="author"><h3 class="author"><span class="firstname">John</span> <span class="surname">Maddock</span></h3></div>
-<div class="author"><h3 class="author"><span class="firstname">Joel</span> <span class="surname">de Guzman</span></h3></div>
-<div class="author"><h3 class="author"><span class="firstname">Eric</span> <span class="surname">Niebler</span></h3></div>
-<div class="author"><h3 class="author"><span class="firstname">Matias</span> <span class="surname">Capeletto</span></h3></div>
-</div>"#;
-    let (_c, _h, a) = detect_copyrights_from_text(text);
-    let names: Vec<&str> = a.iter().map(|d| d.author.as_str()).collect();
-
-    assert!(names.contains(&"John Maddock"), "authors: {names:?}");
-    assert!(names.contains(&"Joel de Guzman"), "authors: {names:?}");
-    assert!(names.contains(&"Eric Niebler"), "authors: {names:?}");
-    assert!(names.contains(&"Matias Capeletto"), "authors: {names:?}");
-}
-
-#[test]
-fn test_detect_created_by_current_user_comment_is_not_author() {
-    let text = "Get the IDs of pipelines created by the current user on the same branch.";
-    let (_c, _h, a) = detect_copyrights_from_text(text);
-    assert!(a.is_empty(), "authors: {a:?}");
 }
 
 #[test]
@@ -2389,68 +1967,6 @@ fn test_detect_copyright_and_author_same_text() {
 }
 
 #[test]
-fn test_detect_author_written_by() {
-    let (_c, _h, a) = detect_copyrights_from_text("Written by Jane Smith");
-    assert_eq!(a.len(), 1, "Should detect one author, got: {:?}", a);
-    assert_eq!(a[0].author, "Jane Smith");
-    assert_eq!(a[0].start_line, LineNumber::ONE);
-    assert_eq!(a[0].end_line, LineNumber::ONE);
-}
-
-#[test]
-fn test_detect_author_maintained_by() {
-    let (_c, _h, a) = detect_copyrights_from_text("Maintained by Bob Jones");
-    assert_eq!(a.len(), 1, "Should detect one author, got: {:?}", a);
-    assert_eq!(a[0].author, "Bob Jones");
-    assert_eq!(a[0].start_line, LineNumber::ONE);
-    assert_eq!(a[0].end_line, LineNumber::ONE);
-}
-
-#[test]
-fn test_detect_author_authors_keyword() {
-    let (_c, _h, a) = detect_copyrights_from_text("Authors John Smith");
-    assert_eq!(
-        a.len(),
-        1,
-        "Should detect author from 'Authors', got: {:?}",
-        a
-    );
-    assert!(
-        a[0].author.contains("John Smith"),
-        "Author: {}",
-        a[0].author
-    );
-}
-
-#[test]
-fn test_detect_author_contributors_keyword() {
-    let (_c, _h, a) = detect_copyrights_from_text("Contributors Jane Doe");
-    assert_eq!(
-        a.len(),
-        1,
-        "Should detect author from 'Contributors', got: {:?}",
-        a
-    );
-    assert!(a[0].author.contains("Jane Doe"), "Author: {}", a[0].author);
-}
-
-#[test]
-fn test_detect_author_spdx_contributor() {
-    let (_c, _h, a) = detect_copyrights_from_text("SPDX-FileContributor: Alice Johnson");
-    assert_eq!(
-        a.len(),
-        1,
-        "Should detect author from SPDX-FileContributor, got: {:?}",
-        a
-    );
-    assert!(
-        a[0].author.contains("Alice Johnson"),
-        "Author: {}",
-        a[0].author
-    );
-}
-
-#[test]
 fn test_detect_copyright_with_company() {
     let (c, h, _a) = detect_copyrights_from_text("Copyright (c) 2024 Google LLC");
     assert_eq!(c.len(), 1, "Should detect one copyright, got: {:?}", c);
@@ -2537,44 +2053,6 @@ fn test_contributed_by_with_utf8_diacritics() {
 }
 
 #[test]
-fn test_name_contributed_line_is_detected_as_author() {
-    let input = "\\author{\nRandall Prium contributed most of the implementation of\n\\code{cut_width()}.\n}";
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Randall Prium"),
-        "expected Randall Prium author, got: {:?}",
-        authors
-    );
-}
-
-#[test]
-fn test_name_contributed_line_ignores_portions_holder_phrase() {
-    let input = "Copyright (c) 2006, Industrial Light & Magic, a division of Lucasfilm\nEntertainment Company Ltd. Portions contributed and copyright held by\nothers as indicated. All rights reserved.";
-    let (_c, _h, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.is_empty(),
-        "expected no authors, got: {:?}",
-        authors
-    );
-}
-
-#[test]
-fn test_date_by_author() {
-    let content = "\
-Copyright (c) 1998 Softweyr LLC.  All rights reserved.
-strtok_r, from Berkeley strtok
-Oct 13, 1998 by Wes Peters <wes@softweyr.com>";
-    let (_c, _h, a) = detect_copyrights_from_text(content);
-    assert!(
-        a.iter().any(|a| a.author.contains("Wes Peters")),
-        "Should detect Wes Peters as author, got: {:?}",
-        a
-    );
-}
-
-#[test]
 fn test_linux_foundation_line_prefers_holder_variant_over_bare_years() {
     let content = "* Copyright (c) 2007, 2010 Linux Foundation";
     let (c, h, _a) = detect_copyrights_from_text(content);
@@ -2610,56 +2088,6 @@ fn test_holder_extracted_from_year_range_with_the_prefix() {
     assert!(
         holders.contains(&"The New York Times Company"),
         "holders: {:?}",
-        h
-    );
-}
-
-#[test]
-fn test_originally_by_author() {
-    let content = "\
-#   Copyright 1996-2006 Free Software Foundation, Inc.
-#   Taken from GNU libtool, 2001
-#   Originally by Gordon Matzigkeit <gord@gnu.ai.mit.edu>, 1996";
-    let (_c, _h, a) = detect_copyrights_from_text(content);
-    assert!(
-        a.iter().any(|a| a.author.contains("Gordon Matzigkeit")),
-        "Should detect Gordon Matzigkeit as author, got: {:?}",
-        a
-    );
-}
-
-#[test]
-fn test_multiline_two_copyrights_adjacent_lines() {
-    let input = "\tCopyright 1988, 1989 by Carnegie Mellon University\n\tCopyright 1989\tTGV, Incorporated\n";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert!(
-        c.iter().any(|cr| cr.copyright.contains("Carnegie Mellon")),
-        "Should detect CMU copyright"
-    );
-    assert!(
-        c.iter().any(|cr| cr.copyright.contains("TGV")),
-        "Should detect TGV copyright, got: {:?}",
-        c
-    );
-    assert!(
-        h.iter().any(|hr| hr.holder.contains("TGV")),
-        "Should detect TGV holder, got: {:?}",
-        h
-    );
-}
-
-#[test]
-fn test_multiline_copyright_after_created_line() {
-    let input = "// Created: Sun Feb  9 10:06:01 2003 by faith@dict.org\n// Copyright 2003, 2004 Rickard E. Faith (faith@dict.org)\n";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert!(
-        c.iter().any(|cr| cr.copyright.contains("Rickard")),
-        "Should detect Faith copyright, got: {:?}",
-        c
-    );
-    assert!(
-        h.iter().any(|hr| hr.holder.contains("Faith")),
-        "Should detect Faith holder, got: {:?}",
         h
     );
 }
@@ -2721,50 +2149,6 @@ fn test_doc_doc_no_overabsorb() {
 }
 
 #[test]
-fn test_multiline_copyrighted_by_href_links_merges_trailing_copyright_clause() {
-    let input = "copyrighted by <A\nHREF=\"http://www.dre.vanderbilt.edu/~schmidt/\">Douglas C. Schmidt</A>\nand his <a\nHREF=\"http://www.cs.wustl.edu/~schmidt/ACE-members.html\">research\ngroup</a> at <A HREF=\"http://www.wustl.edu/\">Washington\nUniversity</A>, <A HREF=\"http://www.uci.edu\">University of California,\nIrvine</A>, and <A HREF=\"http://www.vanderbilt.edu\">Vanderbilt\nUniversity</A>, Copyright (c) 1993-2009, all rights reserved.";
-    let (c, _h, _a) = detect_copyrights_from_text(input);
-    let expected = "copyrighted by http://www.dre.vanderbilt.edu/~schmidt/ Douglas C. Schmidt and his http://www.cs.wustl.edu/~schmidt/ACE-members.html research group at http://www.wustl.edu/ Washington University, http://www.uci.edu University of California, Irvine, and http://www.vanderbilt.edu Vanderbilt University, Copyright (c) 1993-2009";
-    assert!(
-        c.iter().any(|cr| cr.copyright == expected),
-        "Expected merged copyrighted-by href copyright, got: {:?}",
-        c
-    );
-    let merged = c.iter().find(|cr| cr.copyright == expected).unwrap();
-    assert!(
-        merged.end_line > merged.start_line,
-        "Expected merged span to extend across lines, got: {:?}",
-        merged
-    );
-}
-
-#[test]
-fn test_html_anchor_copyright_url_multiline_span_preserved() {
-    let input = "<a href=\"https://example.com/path\">\ncopyright\n</a>";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-
-    let cd = c
-        .iter()
-        .find(|cr| cr.copyright == "copyright https://example.com/path")
-        .unwrap();
-    assert_eq!(
-        (cd.start_line, cd.end_line),
-        (LineNumber::new(1).unwrap(), LineNumber::new(3).unwrap()),
-        "copyrights: {c:?}"
-    );
-
-    let hd = h
-        .iter()
-        .find(|hr| hr.holder == "https://example.com/path")
-        .unwrap();
-    assert_eq!(
-        (hd.start_line, hd.end_line),
-        (LineNumber::new(1).unwrap(), LineNumber::new(3).unwrap()),
-        "holders: {h:?}"
-    );
-}
-
-#[test]
 fn test_json_escaped_html_anchor_copyright_url_detected() {
     let input = r#"&copy; <a href=\"http://www.openstreetmap.org/copyright\">OpenStreetMap</a>"#;
     let (c, h, _a) = detect_copyrights_from_text(input);
@@ -2817,1087 +2201,5 @@ fn test_normalize_split_angle_bracket_urls_keeps_tail() {
         out.contains("https://github.com Havret")
             && out.contains("https://github.com stackbuilders"),
         "normalized: {out:?}"
-    );
-}
-
-#[test]
-fn test_academy_copyright() {
-    let input = "Copyright (c) 2006 Academy of Motion Picture Arts and Sciences";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert!(
-        c.iter()
-            .any(|cr| cr.copyright
-                == "Copyright (c) 2006 Academy of Motion Picture Arts and Sciences"),
-        "Should detect Academy copyright, got: {:?}",
-        c
-    );
-    assert!(
-        h.iter()
-            .any(|hr| hr.holder == "Academy of Motion Picture Arts and Sciences"),
-        "Should detect Academy holder, got: {:?}",
-        h
-    );
-}
-
-#[test]
-fn test_define_copyright() {
-    let input = "#define COPYRIGHT       \"Copyright (c) 1999-2008 LSI Corporation\"\n#define MODULEAUTHOR    \"LSI Corporation\"";
-    let (c, h, a) = detect_copyrights_from_text(input);
-    assert!(
-        c.iter()
-            .any(|cr| cr.copyright == "Copyright (c) 1999-2008 LSI Corporation"),
-        "Should detect 'Copyright (c) 1999-2008 LSI Corporation', got: {:?}",
-        c
-    );
-    assert!(
-        h.iter().any(|h| h.holder == "LSI Corporation"),
-        "Should detect holder, got: {:?}",
-        h
-    );
-    assert!(
-        a.iter().any(|a| a.author == "LSI Corporation"),
-        "Should detect author from MODULEAUTHOR, got: {:?}",
-        a
-    );
-}
-
-#[test]
-fn test_parts_copyright_prefix() {
-    let input = " * Parts (C) 1999 David Airlie, airlied@linux.ie";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert_eq!(c.len(), 1, "expected one copyright, got: {c:#?}");
-    assert_eq!(
-        c[0].copyright, "Parts (c) 1999 David Airlie, airlied@linux.ie",
-        "copyrights: {c:#?}"
-    );
-    assert_eq!(h.len(), 1, "expected one holder, got: {h:#?}");
-    assert_eq!(h[0].holder, "David Airlie");
-}
-
-#[test]
-fn test_trailing_year_included_in_copyright() {
-    let cases = &[
-        (
-            "Copyright (c) IBM Corporation 2008",
-            "Copyright (c) IBM Corporation 2008",
-            "IBM Corporation",
-        ),
-        (
-            "Copyright (c) Zeus Technology Limited 1996",
-            "Copyright (c) Zeus Technology Limited 1996",
-            "Zeus Technology Limited",
-        ),
-        (
-            "Copyright IBM, Corp. 2007",
-            "Copyright IBM, Corp. 2007",
-            "IBM, Corp.",
-        ),
-        (
-            "Copyright IBM Corp. 2004, 2010",
-            "Copyright IBM Corp. 2004, 2010",
-            "IBM Corp.",
-        ),
-    ];
-    for (input, expected_cr, expected_h) in cases {
-        let (c, h, _a) = detect_copyrights_from_text(input);
-        assert!(
-            c.iter().any(|cr| cr.copyright == *expected_cr),
-            "For '{}': expected CR '{}', got {:?}",
-            input,
-            expected_cr,
-            c.iter().map(|cr| &cr.copyright).collect::<Vec<_>>()
-        );
-        assert!(
-            h.iter().any(|hh| hh.holder == *expected_h),
-            "For '{}': expected holder '{}', got {:?}",
-            input,
-            expected_h,
-            h.iter().map(|hh| &hh.holder).collect::<Vec<_>>()
-        );
-    }
-}
-
-#[test]
-fn test_holder_after_year_range_absorbed() {
-    let input = "COPYRIGHT (c) 2006 - 2009 DIONYSOS";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert!(
-        c.iter().any(|cr| cr.copyright.contains("DIONYSOS")),
-        "Should include 'DIONYSOS' in copyright, got: {:?}",
-        c.iter().map(|cr| &cr.copyright).collect::<Vec<_>>()
-    );
-    assert!(
-        h.iter().any(|hh| hh.holder.contains("DIONYSOS")),
-        "Should include 'DIONYSOS' in holder, got: {:?}",
-        h.iter().map(|hh| &hh.holder).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_multi_word_holder_after_year_range() {
-    let input = "Copyright (C) 1999-2000 VA Linux Systems";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert!(
-        c.iter().any(|cr| cr.copyright.contains("VA Linux Systems")),
-        "Should include full company name, got: {:?}",
-        c.iter().map(|cr| &cr.copyright).collect::<Vec<_>>()
-    );
-    assert!(
-        h.iter().any(|hh| hh.holder.contains("VA Linux Systems")),
-        "Should include full company name in holder, got: {:?}",
-        h.iter().map(|hh| &hh.holder).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_by_keyword_holder_captured() {
-    let input = "Copyright (c) 1991, 2000, 2001 by Lucent Technologies.";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert!(
-        c.iter()
-            .any(|cr| cr.copyright.contains("Lucent Technologies")),
-        "Should include holder after 'by', got: {:?}",
-        c.iter().map(|cr| &cr.copyright).collect::<Vec<_>>()
-    );
-    assert!(
-        h.iter().any(|hh| hh.holder.contains("Lucent Technologies")),
-        "Should include holder after 'by', got: {:?}",
-        h.iter().map(|hh| &hh.holder).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_holder_company_with_digits_absorbed() {
-    let input = "Copyright (c) 1995-1996 Guy Eric Schalnat, Group 42, Inc.";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert!(
-        c.iter().any(|cr| cr.copyright.contains("Group 42, Inc.")),
-        "Should include full company name with digits, got: {:?}",
-        c.iter().map(|cr| &cr.copyright).collect::<Vec<_>>()
-    );
-    assert!(
-        h.iter().any(|hh| hh.holder.contains("Group 42, Inc.")),
-        "Should include full company name with digits in holder, got: {:?}",
-        h.iter().map(|hh| &hh.holder).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_copyright_dash_email_tail_absorbed() {
-    let input = "Copyright (c) 1999, Bob Withers - bwit@pobox.com";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert!(
-        c.iter().any(|cr| cr.copyright.contains("bwit@pobox.com")),
-        "Should include dash-email tail in copyright, got: {:?}",
-        c.iter().map(|cr| &cr.copyright).collect::<Vec<_>>()
-    );
-    assert!(
-        h.iter().any(|hh| hh.holder == "Bob Withers"),
-        "Expected holder 'Bob Withers', got: {:?}",
-        h.iter().map(|hh| &hh.holder).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn test_w3c_paren_group_debug() {
-    let input = "(c) 1998-2008 (W3C) MIT, ERCIM, Keio University";
-    let (c, _h, _a) = detect_copyrights_from_text(input);
-    assert!(
-        c.iter()
-            .any(|cr| cr.copyright.contains("MIT, ERCIM, Keio University")),
-        "expected W3C copyright with MIT/ERCIM/Keio, got: {c:#?}"
-    );
-}
-
-#[test]
-fn test_detect_copyright_with_dots_single_line() {
-    // "Copyright . 2008 Foo Name, Inc." - dot after Copyright should be stripped,
-    // and "Foo Name, Inc." should be detected as the full holder.
-    let input = "Copyright . 2008 Foo Name, Inc.";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert_eq!(c.len(), 1, "Should detect one copyright, got: {:?}", c);
-    assert_eq!(
-        c[0].copyright, "Copyright 2008 Foo Name, Inc.",
-        "Should detect full copyright with company name"
-    );
-    assert_eq!(h.len(), 1, "Should detect one holder, got: {:?}", h);
-    assert_eq!(h[0].holder, "Foo Name, Inc.");
-}
-
-#[test]
-fn test_detect_copyright_with_dots_multiline() {
-    let input = "Copyright . 2008 company name, inc.";
-    let (c, h, _a) = detect_copyrights_from_text(input);
-    assert!(
-        !c.is_empty(),
-        "Should detect at least one copyright, got: {:?}",
-        c
-    );
-    assert!(
-        c.iter().any(|cr| cr.copyright.contains("2008")),
-        "Should detect copyright with year 2008, got: {:?}",
-        c
-    );
-    assert!(
-        c.iter()
-            .any(|cr| cr.copyright.to_lowercase().contains("company name")),
-        "Should detect full company name, got: {:?}",
-        c
-    );
-    assert!(
-        h.iter()
-            .any(|hr| hr.holder.to_lowercase().contains("company name")),
-        "Should detect holder with company name, got: {:?}",
-        h
-    );
-}
-
-#[test]
-fn test_opensharedmap_inc_holder_detected() {
-    let input = "Copyright (C) OpenSharedMap Inc.";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.copyright.contains("OpenSharedMap Inc")),
-        "Expected OpenSharedMap copyright detection, got: {copyrights:?}"
-    );
-    assert!(
-        holders
-            .iter()
-            .any(|h| h.holder.contains("OpenSharedMap Inc")),
-        "Expected OpenSharedMap holder detection, got: {holders:?}"
-    );
-}
-
-#[test]
-fn test_disclaimer_tail_with_inc_as_does_not_create_holder() {
-    let input = "Copyright Owner Inc. AS IS AND ANY EXPRESS OR IMPLIED WARRANTIES";
-    let (_copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        holders.is_empty(),
-        "Unexpected disclaimer-derived holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_camelcase_provider_not_author_false_positive() {
-    let input = "A meter implementation is created by a MeterProvider in this system.\nA trace implementation is created by a TracerProvider in this system.";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    let author_values: Vec<&str> = authors.iter().map(|a| a.author.as_str()).collect();
-    assert!(
-        author_values
-            .iter()
-            .all(|a| *a != "MeterProvider in" && *a != "TracerProvider in"),
-        "Unexpected provider false-positive authors: {author_values:?}"
-    );
-}
-
-#[test]
-fn test_platformdirs_lowercase_holder_detected() {
-    let input = "Copyright (c) 2010-202x The platformdirs developers";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights.iter().any(|c| c
-            .copyright
-            .contains("2010-202x The platformdirs developers")),
-        "Expected platformdirs copyright, got: {copyrights:?}"
-    );
-    assert!(
-        holders
-            .iter()
-            .any(|h| h.holder == "The platformdirs developers"),
-        "Expected platformdirs holder, got: {holders:?}"
-    );
-}
-#[test]
-fn test_square_c_sign_detected() {
-    let input = "[C] The Regents of the University of Michigan and Merit Network, Inc. 1992, 1993, 1994, 1995 All Rights Reserved";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights.iter().any(|c| c
-            .copyright
-            .contains("Regents of the University of Michigan")),
-        "Expected Regents copyright detection, got: {copyrights:?}"
-    );
-    assert!(
-        holders
-            .iter()
-            .any(|h| h.holder.contains("Regents of the University of Michigan")),
-        "Expected Regents holder detection, got: {holders:?}"
-    );
-}
-
-#[test]
-fn test_template_literal_copyright_holder_detected() {
-    let input = "copyright: `Copyright 2010–${new Date().getUTCFullYear()} Mike Bostock`";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.copyright == "Copyright 2010-${new Date .getUTCFullYear } Mike Bostock"),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        holders.iter().any(|h| h.holder == "Mike Bostock"),
-        "holders: {holders:?}"
-    );
-    assert!(
-        !copyrights.iter().any(|c| c.copyright == "Copyright 2010-$"),
-        "copyrights: {copyrights:?}"
-    );
-}
-
-#[test]
-fn test_markdown_transition_line_not_author() {
-    let input = "The meaning of [*transition*.delay](https://github.com/d3/d3-transition/blob/master/README.md#transition_delay) has changed for chained transitions created by [*transition*.transition](https://github.com/d3/d3-transition/blob/master/README.md#transition_transition).";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        !authors.iter().any(|a| a.author.contains(
-            "transition .transition https://github.com/d3/d3-transition/blob/master/README.md"
-        )),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_json_author_field_does_not_capture_following_metadata_blob() {
-    let input = "author: Box UK,\nurl: http://updates.jenkins-ci.org/download/plugins/jslint/0.7.6/jslint.hpi,\nversion: 0.7.6,\nwiki: https://wiki.jenkins-ci.org/display/JENKINS/JSLint+plugin";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    let author_values: Vec<&str> = authors.iter().map(|a| a.author.as_str()).collect();
-    assert_eq!(author_values, vec!["Box UK"], "authors: {authors:?}");
-}
-
-#[test]
-fn test_sentence_fragment_not_author() {
-    let input = "with key equal to unescaped token";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_call_ref_fragment_not_author() {
-    let input = "call @ref";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_boost_value_stack_call_ref_sentence_not_author() {
-    let input = "Then to build a @ref value, first call @ref reset and optionally specify the boost::container::pmr::memory_resource.";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_json_excerpt_developed_by_company_author_detected() {
-    let input = r#"mes/0.2/jsgames.hpi","version":"0.2","wiki":"https://wiki.jenkins-ci.org/display/JENKINS/JSGames+Plugin"},"jslint":{"buildDate":"Jan 03, 2013","dependencies":[],"developers":[{"developerId":"gavd","email":"gavin.davies@boxuk.com","name":"Gavin Davies"}],"excerpt":"Lint JavaScript files, outputting to checkstyle format. Supports all JSLint options. Developed by Box UK.","gav":"org.jenkins-ci.plugins:jslint:0.7.6","labels":["misc"],"name":"jslint","previousTimestamp":"2013-01-03T20:22:38.00Z","previousVersion":"0.7.5","releaseTimestamp":"2013-01-03T20:29:06.00Z","requiredCore":"1.474","scm":"github.com""#;
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Box UK"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_plain_json_author_string_preserved() {
-    let input = r#""author": "Google's Web DevRel Team","#;
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert_eq!(
-        authors
-            .iter()
-            .map(|a| a.author.as_str())
-            .collect::<Vec<_>>(),
-        vec!["Google's Web DevRel Team"],
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_plain_json_author_string_with_parenthesized_url_preserved() {
-    let input = r#""author": "Qix (http://github.com/qix-)","#;
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert_eq!(
-        authors
-            .iter()
-            .map(|a| a.author.as_str())
-            .collect::<Vec<_>>(),
-        vec!["Qix (http://github.com/qix-)"],
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_plain_json_author_string_with_parenthesized_url_and_following_key_preserved() {
-    let input = concat!(
-        "   \"author\": \"Qix (http://github.com/qix-)\",\n",
-        "   \"keywords\": [\n",
-    );
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert_eq!(
-        authors
-            .iter()
-            .map(|a| a.author.as_str())
-            .collect::<Vec<_>>(),
-        vec!["Qix (http://github.com/qix-)"],
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_json_author_object_name_preferred_over_url_tail() {
-    let input =
-        "\"author\": { \"name\": \"Chen Fengyuan\", \"url\": \"https://chenfengyuan.com/\" }";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert_eq!(
-        authors
-            .iter()
-            .map(|a| a.author.as_str())
-            .collect::<Vec<_>>(),
-        vec!["Chen Fengyuan"],
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_multiline_json_author_object_name_detected() {
-    let input = concat!(
-        "  \"author\": {\n",
-        "    \"name\": \"Chen Fengyuan\",\n",
-        "    \"url\": \"https://chenfengyuan.com/\"\n",
-        "  }\n",
-    );
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors.iter().any(|a| a.author == "Chen Fengyuan"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_metadata_json_author_fallback_keeps_collective_and_single_word_names() {
-    let input = r#"{
-  "components": [
-    {
-      "supplier": { "name": "Google LLC" },
-      "author": "gRPC authors",
-      "name": "gRPC (C++)"
-    },
-    {
-      "supplier": { "name": "Meta Open Source" },
-      "author": "Meta",
-      "name": "folly"
-    },
-    {
-      "supplier": { "name": "The libunwind project" },
-      "author": "The libunwind project",
-      "name": "libunwind"
-    },
-    {
-      "supplier": { "name": "Google LLC" },
-      "author": "S2Geometry",
-      "name": "S2 Geometry Library"
-    }
-  ]
-}"#;
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    let values: Vec<&str> = authors.iter().map(|a| a.author.as_str()).collect();
-    assert!(values.contains(&"gRPC authors"), "authors: {authors:?}");
-    assert!(values.contains(&"Meta"), "authors: {authors:?}");
-    assert!(
-        values.contains(&"The libunwind project"),
-        "authors: {authors:?}"
-    );
-    assert!(values.contains(&"S2Geometry"), "authors: {authors:?}");
-}
-
-#[test]
-fn test_json_code_example_author_fields_do_not_create_authors() {
-    let input = r#"{
-  "expectedStages": [
-    {
-      "$match": {
-        "author": "Agatha Christie"
-      }
-    },
-    {
-      "$setMetadata": {
-        "score": {
-          "$divide": [1, 2]
-        }
-      }
-    }
-  ]
-}"#;
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_code_pipeline_author_match_not_detected_as_author() {
-    let input = r#"{
-  $scoreFusion: {
-    input: {
-      pipelines: {
-        pipeOne: [
-          { $match : { author : "Agatha Christie" } },
-          { $sort: {author: 1} }
-        ]
-      }
-    }
-  }
-}"#;
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_json_sponsor_description_does_not_create_authors() {
-    let input = r#"{
-  "description": "A useful plugin",
-  "sponsor": {
-    "@type": "Organization",
-    "name": "Example Org",
-    "description": "Developer as a service Plugin in the cloud when a workflow runs."
-  }
-}"#;
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_fast_path_proposal_phrase_not_author() {
-    let input = "Clinger's fast path, inspired by Jakub Jelínek's proposal";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_readme_security_review_prose_not_author() {
-    let input = "application developers can trust, the C++ Alliance has commissioned Bishop Fox to perform a security audit of the Boost.JSON library. The report linked here";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_tomcat_html_doc_prose_not_author() {
-    let input = "the order defined by the DTD (see Section 13.3).</p>";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_tomcat_contributing_prose_not_author() {
-    let input = "time as all committers are volunteers on the project. If a significant amount";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_tomcat_footer_trademark_line_not_absorbed_into_copyright() {
-    let input = "Copyright (c) 1999-2026, The Apache Software Foundation\nApache Tomcat, Tomcat, Apache, the Apache Tomcat logo and the Apache logo\nare either registered trademarks or trademarks of the Apache Software Foundation.";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.copyright == "Copyright (c) 1999-2026, The Apache Software Foundation"),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        !copyrights.iter().any(|c| c
-            .copyright
-            .contains("Apache Tomcat, Tomcat, Apache, the Apache Tomcat")),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        holders
-            .iter()
-            .any(|h| h.holder == "The Apache Software Foundation"),
-        "holders: {holders:?}"
-    );
-    assert!(
-        !holders.iter().any(|h| h
-            .holder
-            .contains("Apache Tomcat, Tomcat, Apache, the Apache Tomcat")),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_update_center_metadata_blob_not_multiple_authors() {
-    let input = "author: Box UK, url: http://updates.jenkins-ci.org/download/plugins/jslint/0.7.6/jslint.hpi, version: 0.7.6, wiki: https://wiki.jenkins-ci.org/display/JENKINS/JSLint+plugin, title: JSLint plugin, buildDate: Jan 03, 2013, developerId: gavd";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    let values: Vec<&str> = authors.iter().map(|a| a.author.as_str()).collect();
-    assert_eq!(values, vec!["Box UK"], "authors: {authors:?}");
-}
-
-#[test]
-fn test_gsoc_javascript_language_phrase_not_author() {
-    let input = "My proposal is based on getting full support for JavaScript within the RoboComp framework. For this, the current state of generation of written components in the JavaScript language must be improved.";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        !authors
-            .iter()
-            .any(|a| a.author == "components in the JavaScript language"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_apostrophized_person_name_metadata_not_author() {
-    let input = "type' Person name' AadityaNair";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_language_fragment_not_author() {
-    let input = "in PHP";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_single_word_stopword_not_author() {
-    let input = "In";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_short_comma_metadata_phrase_not_author() {
-    let input = "GENIVI, several standard";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_short_url_suffix_phrase_not_author() {
-    let input = "J.L.Blanco (https://github.com/jlblancoc)";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_lowercase_starting_multiword_fragment_not_author() {
-    let input = "around the world. It";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_dangling_two_word_phrase_not_author() {
-    let input = "Sandcastle and";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_gsoc_spdx_sentence_not_copyright_or_holder() {
-    let input = "Software Package Data Exchange (SPDX) is a set of standards for communicating the components, licenses, and copyrights associated with software.";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(copyrights.is_empty(), "copyrights: {copyrights:?}");
-    assert!(holders.is_empty(), "holders: {holders:?}");
-}
-
-#[test]
-fn test_json_description_and_sponsor_not_copyright_or_holder() {
-    let input = r#""description": "Software Package Data Exchange (SPDX) is a set of standards for communicating the components, licenses, and copyrights associated with software.", "sponsor": { "@type": "Organization", "name": "FOSSology", "disambiguatingDescription": "Open Source License Compliance by Open Source Software", "url": "http://example.com/logo" }"#;
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(copyrights.is_empty(), "copyrights: {copyrights:?}");
-    assert!(holders.is_empty(), "holders: {holders:?}");
-}
-
-#[test]
-fn test_mit_intro_not_copyright_or_holder() {
-    let input = "These files are covered by the following copyright and MIT License, reproduced from the original project:";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        !copyrights
-            .iter()
-            .any(|c| c.copyright == "copyright and MIT"),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        !holders.iter().any(|h| h.holder == "MIT"),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_rest_api_description_not_holder_or_copyright() {
-    let input = "We provide developers, researchers, and students the ability to access any model using a simple REST API call. The REST API description.";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(copyrights.is_empty(), "copyrights: {copyrights:?}");
-    assert!(holders.is_empty(), "holders: {holders:?}");
-}
-
-#[test]
-fn test_bare_rest_marker_not_copyright() {
-    let input = "(c) REST";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(copyrights.is_empty(), "copyrights: {copyrights:?}");
-    assert!(holders.is_empty(), "holders: {holders:?}");
-}
-
-#[test]
-fn test_semicolon_joined_copyright_list_does_not_keep_combined_variant() {
-    let input =
-        "(C) 2019--2020 Vinnie Falco; (C) 2020 Krystian Stasiowski; (C) 2022 Dmitry Arkhipov";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        !copyrights.iter().any(|c| {
-            c
-            .copyright
-            == "(c) 2019-2020 Vinnie Falco; (c) 2020 Krystian Stasiowski; (c) 2022 Dmitry Arkhipov"
-        }),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.copyright == "(c) 2019-2020 Vinnie Falco"),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.copyright == "(c) 2020 Krystian Stasiowski"),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.copyright == "(c) 2022 Dmitry Arkhipov"),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        holders.iter().any(|h| h.holder == "Vinnie Falco")
-            && holders.iter().any(|h| h.holder == "Krystian Stasiowski")
-            && holders.iter().any(|h| h.holder == "Dmitry Arkhipov"),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_normalize_company_suffix_period_holder_variants() {
-    let input = "Copyright (c) 2020 Foo, Inc\nCopyright (c) 2021 Foo, Inc.";
-    let (_copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert_eq!(holders.len(), 2, "holders: {holders:?}");
-    assert!(
-        holders.iter().all(|h| h.holder == "Foo, Inc."),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_split_multiline_holder_list_with_emails() {
-    let input = "(c) 1999                Terrehon Bowden <terrehon@pacbell.net>\n                        Bodo Bauer <bb@ricochet.net>\n";
-
-    let (_copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        holders.iter().any(|h| h.holder == "Terrehon Bowden"),
-        "holders: {holders:?}"
-    );
-    assert!(
-        holders.iter().any(|h| h.holder == "Bodo Bauer"),
-        "holders: {holders:?}"
-    );
-    assert!(
-        !holders
-            .iter()
-            .any(|h| h.holder == "Terrehon Bowden Bodo Bauer"),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_boost_style_multiline_holder_continuation_after_year_first_line() {
-    let input = "// Copyright (c) 2019 Peter Dimov (pdimov at gmail dot com),\n\
-//                    Vinnie Falco (vinnie.falco@gmail.com)\n\
-// Copyright (c) 2020 Krystian Stasiowski (sdkrystian@gmail.com)\n";
-
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights.iter().any(|c| {
-            c.start_line == LineNumber::ONE
-                && c.end_line == LineNumber::new(2).unwrap()
-                && c.copyright.contains("Peter Dimov")
-                && c.copyright.contains("Vinnie Falco")
-        }),
-        "copyrights: {copyrights:?}"
-    );
-
-    assert!(
-        holders.iter().any(|h| {
-            h.start_line == LineNumber::ONE
-                && h.end_line == LineNumber::new(2).unwrap()
-                && h.holder.contains("Peter Dimov")
-                && h.holder.contains("Vinnie Falco")
-        }),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_year_first_multiline_holder_repair_does_not_absorb_multiline_holder_lists() {
-    let input = "Copyright (c) 1995, 1996, 1997 Francis.Dupont@inria.fr, INRIA Rocquencourt,\n\
-Alain.Durand@imag.fr, IMAG,\n\
-Jean-Luc.Richier@imag.fr, IMAG-LSR.\n";
-
-    let (copyrights, _holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights.iter().any(|c| {
-            c.start_line == LineNumber::ONE
-                && c.end_line == LineNumber::ONE
-                && c.copyright == "Copyright (c) 1995, 1996, 1997 Francis.Dupont@inria.fr, INRIA"
-        }),
-        "copyrights: {copyrights:?}"
-    );
-
-    assert!(
-        !copyrights
-            .iter()
-            .any(|c| c.copyright.contains("Rocquencourt") || c.copyright.contains("Alain.Durand")),
-        "copyrights: {copyrights:?}"
-    );
-}
-
-#[test]
-fn test_extend_copyright_with_following_all_rights_reserved_line() {
-    let input = "Copyright 2010-2015 Mike Bostock\nAll rights reserved.";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.copyright == "Copyright 2010-2015 Mike Bostock"),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.start_line == LineNumber::ONE && c.end_line == LineNumber::new(2).unwrap()),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        holders.iter().any(|h| h.holder == "Mike Bostock"),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_drop_url_embedded_suffix_copyright_and_holder_variants() {
-    let input =
-        "/* Copyright (c) 2020 Example Corp. See url(\"https://dummy-url-for-test.com\") */";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.copyright == "Copyright (c) 2020 Example Corp."),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        !copyrights
-            .iter()
-            .any(|c| c.copyright.contains("See url") || c.copyright.contains("https://")),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        holders.iter().any(|h| h.holder == "Example Corp."),
-        "holders: {holders:?}"
-    );
-    assert!(
-        !holders
-            .iter()
-            .any(|h| h.holder.contains("See url") || h.holder.contains("http")),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_add_missing_holder_from_preceding_name_line_for_year_only_copyright() {
-    let input = "Author:  David Beazley (http://www.dabeaz.com)\nCopyright (C) 2007\n";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.copyright == "David Beazley, Copyright (c) 2007"),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        holders.iter().any(|h| h.holder == "David Beazley"),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_descriptive_line_does_not_expand_year_only_copyright_holder() {
-    let input = "Tru64 audio module for SDL (Simple DirectMedia Layer)\nCopyright (C) 2003\n";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights
-            .iter()
-            .any(|c| c.copyright == "Copyright (c) 2003"),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        !copyrights
-            .iter()
-            .any(|c| c.copyright == "Tru64 audio module for SDL, Copyright (c) 2003"),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        holders
-            .iter()
-            .all(|h| h.holder != "Tru64 audio module for SDL"),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_drop_trademarked_materials_prose_false_positive_copyrights_and_holders() {
-    let input = "SPDX-FileCopyrightText: <years> Univention GmbH\n\nBinary versions of this program provided by Univention to you as well\nas other copyrighted, protected or trademarked materials like Logos,\ngraphics, fonts, specific documentations and configurations,\ncryptographic keys etc. are subject to a license agreement between you\nand Univention and not subject to the AGPL-3.0-only.\n";
-    let (copyrights, holders, _authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        copyrights
-            .iter()
-            .all(|c| !c.copyright.contains("protected or trademarked materials")),
-        "copyrights: {copyrights:?}"
-    );
-    assert!(
-        holders
-            .iter()
-            .all(|h| !h.holder.contains("protected or trademarked materials")),
-        "holders: {holders:?}"
-    );
-    assert!(
-        holders.iter().any(|h| h.holder == "Univention GmbH"),
-        "holders: {holders:?}"
-    );
-}
-
-#[test]
-fn test_created_by_without_handle_or_email_is_rejected() {
-    let (_copyrights, _holders, authors) =
-        detect_copyrights_from_text("Created by IntelliJ IDEA\n");
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_created_by_tool_banner_is_rejected() {
-    let (_copyrights, _holders, authors) =
-        detect_copyrights_from_text("created by Grunt and NPM.\n");
-    assert!(authors.is_empty(), "authors: {authors:?}");
-}
-
-#[test]
-fn test_extract_toml_authors_array_as_single_combined_detection() {
-    let input = "authors = [\"The Rand Project Developers\", \"The Rust Project Developers\"]\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "The Rand Project Developers The Rust Project Developers"),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_normalize_original_author_current_tail() {
-    let input = "* Original author: Chris Pallotta <chris@allmedia.com>\n* Current maintainer: Jim Van Zandt <jrv@vanzandt.mv.com>\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Chris Pallotta <chris@allmedia.com>"),
-        "authors: {authors:?}"
-    );
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Jim Van Zandt <jrv@vanzandt.mv.com>"),
-        "authors: {authors:?}"
-    );
-    assert!(
-        !authors.iter().any(|a| a.author.contains("Current")),
-        "authors: {authors:?}"
-    );
-}
-
-#[test]
-fn test_normalize_original_authors_multiline_to_separate_people() {
-    let input = "* Original Authors: Robert Jennings <rcj@linux.vnet.ibm.com>\n*                   Seth Jennings <sjenning@linux.vnet.ibm.com>\n";
-    let (_copyrights, _holders, authors) = detect_copyrights_from_text(input);
-
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Robert Jennings <rcj@linux.vnet.ibm.com>"),
-        "authors: {authors:?}"
-    );
-    assert!(
-        authors
-            .iter()
-            .any(|a| a.author == "Seth Jennings <sjenning@linux.vnet.ibm.com>"),
-        "authors: {authors:?}"
-    );
-    assert!(
-        !authors.iter().any(|a| {
-            a.author
-                == "Robert Jennings <rcj@linux.vnet.ibm.com> Seth Jennings <sjenning@linux.vnet.ibm.com>"
-        }),
-        "authors: {authors:?}"
     );
 }
